:art: move check_key_len to Client
